requirements.txt
requests
openai
python-dotenv
schedule
mysql-connector-python